Desconstructed Preferences app to helper files
