+++ src/components/grommet/toggle.js
@@ -1,0 +1,58 @@
+import { css } from "styled-components";
+
+
+const checkboxCheckStyle = css`
+  background-color: #FBDCA1;
+  border-color: #FBDCA1;
+`;
+
+const customToggleTheme = {
+  global: {
+    colors: {
+      "focus": undefined,
+      "toggle-bg": "#CCCCCC",
+      "toggle-knob": "white",
+      "toggle-accent": "#CCCCCC"
+    }
+  },
+  checkBox: {
+    border: {
+      color: {
+        light: "toggle-bg"
+      }
+    },
+    color: {
+      light: "toggle-knob"
+    },
+    check: {
+      radius: "2px"
+    },
+    hover: {
+      border: {
+        color: undefined
+      }
+    },
+    toggle: {
+      background: { light: "toggle-accent" },
+      color: {
+        light: "toggle-knob"
+      },
+      size: "35px",
+      knob: {
+        extend: `
+          top: 0px;
+          box-shadow: 0px 0px 2px 0px rgba(66,66,66,1),
+           0px 2px 2px 0px rgba(66,66,66,1);
+        `
+      },
+      extend: ({ checked }) => `
+        height: 22px;
+        ${checked && checkboxCheckStyle}
+      `
+    },
+    gap: "xsmall",
+    size: "18px"
+  }
+};
+
+export default customToggleTheme;
